+++ test/unit/fuzz-tools.test.js
@@ -166,6 +166,7 @@ import * as pkV166 from '../../lib/pk-v166.js';
 import * as oneformulaV167 from '../../lib/oneformula-v167.js';
 import * as ltcgaV173 from '../../lib/ltcga-v173.js';
 import * as ltcgaV174 from '../../lib/ltcga-v174.js';
+import * as ltcgaV175 from '../../lib/ltcga-v175.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -276,6 +277,7 @@ const MODULES = {
   'oneformula-v167.js': oneformulaV167,
   'ltcga-v173.js': ltcgaV173,
   'ltcga-v174.js': ltcgaV174,
+  'ltcga-v175.js': ltcgaV175,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];